Move shared useful platform utils into SDK and update plugin-path code
