+++ backend/src/platform_utils.linux.cpp
@@ -8,6 +8,8 @@
 #include "hipdnn_exception.hpp"
 #include <dlfcn.h>
 
+using hipdnn_sdk::utilities::Plugin_lib_handle;
+
 namespace hipdnn_backend::platform_utils
 {
