+++ backend/tests/test_shared_library.cpp
@@ -21,7 +21,7 @@ const char* const SYMBOL_NAME = "hipdnnPluginGetName";
 const char* const WRONG_SYMBOL_NAME = "wrong_symbol_name";
 
 const std::string FULL_LIBRARY_PATH = std::filesystem::path(".")
-    /= platform_utils::get_library_name("hipdnn_test_plugin1");
+    /= hipdnn_sdk::utilities::get_library_name("hipdnn_test_plugin1");
 
 }
